server/healthaggregation/requirements/prod.txt
# Everything needed in production

setuptools==57.0.0

# Flask
Flask==2.0.2
Werkzeug==2.0.2
click>=7.0
flask-jwt-extended==4.3.1

# Database
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.27
psycopg2-binary==2.9.2

# Migrations
Flask-Migrate==3.1.0

# Forms
email-validator==1.1.3
Flask-WTF==1.0.0
WTForms==3.0.0

# Deployment
gevent==21.8.0
gunicorn>=19.9.0
supervisor==4.2.2

# Flask Static Digest
Flask-Static-Digest==0.2.1

# Auth
Flask-Login==0.5.0
Flask-Bcrypt==0.7.1

# Caching
Flask-Caching>=1.7.2

# Debug toolbar
Flask-DebugToolbar==0.11.0

# Environment variable parsing
environs==9.3.5

# SSL
pyopenssl==21.0.0

# Firebase Cloud Messaging
pyfcm==1.5.4

# Pseudorandom Number Generator / LibSodium Binding
pynacl~=1.5.0

# Array Calculations
numpy~=1.21.4

# Cryptography
GF256~=0.2.0